+++ configs/species_info.json
