+++ fluxy/PARIS_inversion_results.py
@@ -1,0 +1,85 @@
+import xarray as xr
+import numpy as np
+from pandas import to_datetime
+import matplotlib.pyplot as plt
+import os
+import glob
+import math
+from matplotlib.dates import YearLocator, MonthLocator
+from matplotlib.ticker import NullFormatter, AutoMinorLocator, MultipleLocator
+import pprint
+import cartopy
+
+import inspect
+from IPython.utils import io
+import sys
+
+
+
+#####################################################################
+
+
+#####################################################################
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+#####################################################################
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+#####################################################################
+
+
+#####################################################################
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+
+#####################################################################
+
+#####################################################################
+
+
+
+
+
