+++ fluxy/plots/mf_stats.py
@@ -1,0 +1,89 @@
+import matplotlib.pyplot as plt
+
+def plot_stats_mf(pearson,nrmse,rmse,species,
+                  model_colors,s_data,m_data,
+                  start_date=None,end_date=None):
+    """
+    Plots fit statistics for all sites, for all models.
+    
+    Args:
+        pearson (dictionary of dictionaries):
+            Pearson correlation coeffiecient, for each site and for each model.
+        nrmse (dictionary of dictionaries):
+            Normalised root mean square error, for each site and for each model.
+        rmse (dictionary of dictionaries):
+            Root mean square error, for each site and for each model (not normalized).
+        species (str): 
+            Gas species, e.g. 'ch4'.
+        model_colors (dict of str):
+            Models and corresponding colours used to plot the model.
+        s_data (dict of dict):
+            Dictionary of species with information for plotting (read from json file).
+        m_data (dict of dict):
+            Dictionary of inversion runs with filename and plot label (read from json file).
+        start_date (str) and end_date (str):
+            Dates used to title the plot. 
+    Returns:
+        fig (figure): 
+            Three plots showing each model's fit statistics, for each site.
+    """
+    
+    x_val = []
+    x_label = []
+
+    model_colors_stats = {'intem':'dodgerblue',
+                        'elris_name':'purple'}
+
+    fig,ax = plt.subplots(3,1,figsize=(10,9),tight_layout=True)
+    
+    for i,site in enumerate(pearson.keys()):
+        for m,model in enumerate(pearson[site]):
+            model0 = model.split('_')[0]
+            if i == 0:
+                ax[0].scatter(i+m*0.2,pearson[site][model],color=model_colors[model][0],marker='x',s=150,label=m_data[model]["label"])
+                ax[1].scatter(i+m*0.2,nrmse[site][model],color=model_colors[model][0],marker='x',s=150,label=m_data[model]["label"])
+                ax[2].scatter(i+m*0.2,rmse[site][model],color=model_colors[model][0],marker='x',s=150,label=m_data[model]["label"])
+                #ax[2].scatter(i+m*0.2,std[site][model],color=model_colors_stats[model],marker='x',s=150,label=m_data[model]["label"])
+                
+            else:
+                ax[0].scatter(i+m*0.2,pearson[site][model],color=model_colors[model][0],marker='x',s=150)
+                ax[1].scatter(i+m*0.2,nrmse[site][model],color=model_colors[model][0],marker='x',s=150)
+                ax[2].scatter(i+m*0.2,rmse[site][model],color=model_colors[model][0],marker='x',s=150)
+                #ax[2].scatter(i+m*0.2,std[site][model],color=model_colors_stats[model],marker='x',s=150)
+                
+        x_val.append(i)
+        x_label.append(site)
+        
+    #y_lim0 = [ax[0].get_ylim()[0],ax[0].get_ylim()[1]]
+    #y_lim1 = [ax[0].get_ylim()[0],ax[0].get_ylim()[1]]
+    
+    for i in range(3):
+        ax[i].set_xticks(x_val);
+        ax[i].set_xticklabels(x_label,rotation=45);
+        ax[i].set_xlim(x_val[0]-0.2,x_val[-1]+0.4)
+        #y_lim = [ax[i].get_ylim()[0],ax[i].get_ylim()[1]]
+        #ax[i].set_ylim([y_lim[0]-0.1*y_lim[0],y_lim[1]+y_lim[1]*0.1])
+                
+    ax[0].invert_yaxis()
+    ax[0].hlines(1,x_val[0]-0.2,x_val[-1]+0.4,linestyle='dotted',color='grey')        
+    ax[1].hlines(0,x_val[0]-0.2,x_val[-1]+0.4,linestyle='dotted',color='grey')        
+    ax[2].hlines(0,x_val[0]-0.2,x_val[-1]+0.4,linestyle='dotted',color='grey')
+    
+    ax[0].set_ylabel('Pearson\n correlation coefficient')
+    ax[1].set_ylabel('Normalised RMSE')
+    ax[2].set_ylabel('RMSE')
+    #ax[2].set_ylabel('Standard\ndeviation')
+
+    leg = ax[0].legend(ncol=2,borderpad=.2,columnspacing=1.0)
+    try:
+        for l in leg.legend_handles:
+            l.set_linewidth(5.0)
+    except:
+        for l in leg.legendHandles:
+            l.set_linewidth(5.0)
+
+    fig.suptitle((f'{s_data[species]["species_print"]} Modelled mole fraction statistical fit to obs')+
+                 f' \n{start_date} to {end_date}')
+    
+    
+    return fig
